fix PR issues
