+++ server/index.js
@@ -18,7 +18,7 @@ import whisperRoutes from "./routes/whisperRoutes.js";
 import gptSemanticSearchRoutes from "./routes/gptSemanticSearchRoutes.js";
 import authRoutes from "./routes/authRoutes.js";
 import userRoutes from "./routes/userRoutes.js";
-import { loadTrainingData } from "./utils.js";
+import { loadTrainingData } from "./services/pinecone.js";
 import { initSocketIO } from "./socketIO.js";
 
 const app = express();
@@ -62,20 +62,14 @@ app.use((req, res, next) => {
   next();
 });
 
+const port = process.env.PORT || 8080;
+
 const startServer = async () => {
   try {
     connectDB(process.env.MONGO_DB_URL);
-
-    if (process.env.NODE_ENV === "development") {
-      server.listen(8080, () =>
-        console.log(`Server started on port 8080, in development mode`)
-      );
-    } else {
-      console.log("production");
-      server.listen(process.env.PORT, () =>
-        console.log(`Server started on port ${process.env.PORT}`)
-      );
-    }
+    server.listen(port, () =>
+      console.log(`Server started on port ${port}, in development mode`)
+    );
   } catch (error) {
     console.log(error);
   }
